+++ src/main/groovy/com/bmuschko/gradle/docker/tasks/container/extras/DockerLivenessContainer.groovy
@@ -16,6 +16,10 @@
 
 package com.bmuschko.gradle.docker.tasks.container.extras
 
+import org.gradle.api.tasks.Internal
+
+import java.util.concurrent.TimeUnit
+
 import static com.bmuschko.gradle.docker.utils.IOUtils.getProgressLogger
 
 import com.bmuschko.gradle.docker.domain.LivenessProbe
@@ -26,64 +30,76 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Optional
 
 /**
- *  Poll a given running container for an arbitrary log message to confirm liveness. If a probe is
+ *  Poll a given running container for an arbitrary log message to confirm liveness. If a livenessProbe is
  *  NOT defined then we fallback to check if the container is in a running state.
  */
-class DockerLivenessProbeContainer extends DockerLogsContainer {
+class DockerLivenessContainer extends DockerLogsContainer {
 
     @Input
     @Optional
-    LivenessProbe probe
+    LivenessProbe livenessProbe
+
+    // last call to inspect container which will only
+    // be non-null once task has completed execution.
+    @Internal
+    private lastInspection
 
     @Override
     void runRemoteCommand(dockerClient) {
         logger.quiet "Starting liveness probe on container with ID '${getContainerId()}'."
 
-        // if probe was defined proceed as expected otherwise just
+        // if livenessProbe was defined proceed as expected otherwise just
         // check if the container is up and running
-        if (probe) {
-
-            // if not already set the following will be defaulted
-            if (!this.getSince()) {
-                this.setSince(new Date())
-            }
-            if (!this.getTailCount()) {
-                this.setTailCount(10)
-            }
+        if (livenessProbe) {
 
             // create progressLogger for pretty printing of terminal log progression
-            final def progressLogger = getProgressLogger(project, DockerLivenessProbeContainer)
+            final def progressLogger = getProgressLogger(project, DockerLivenessContainer)
             progressLogger.started()
 
             boolean matchFound = false
-            long localPollTime = probe.pollTime
+            long localPollTime = livenessProbe.pollTime
             int pollTimes = 0
 
             // 1.) Write the content of the logs into a StringWriter which we zero-out
             //     below after each successive log grab.
             setSink(new StringWriter())
+            setTailCount(null)
+            setTailAll(null)
+            setFollow(null)
+
+            Date lastDate = this.getSince()
 
             while (localPollTime > 0) {
-                pollTimes = pollTimes + 1
+                pollTimes += 1
 
                 // 2.) check if container is actually running
-                def container = dockerClient.inspectContainerCmd(getContainerId()).exec()
-                if (container.getState().getRunning() == false) {
+                lastInspection = dockerClient.inspectContainerCmd(getContainerId()).exec()
+                if (lastInspection.getState().getRunning() == false) {
                     throw new GradleException("Container with ID '${getContainerId()}' is not running and so can't perform liveness probe.");
                 }
+
+                if (lastDate) {
+                    this.setSince(lastDate)
+                }
+                final Date justBeforeLastExecutionDate = Calendar.getInstance().getTime()
 
                 // 3.) execute our "special" version of `runRemoteCommand` to
                 //     check if next log line has the message we're interested in
                 //     which in turn will have its output written into the sink.
                 _runRemoteCommand(dockerClient)
 
+                lastDate = justBeforeLastExecutionDate
+
                 // 4.) check if log contains expected message otherwise sleep
                 String logLine = getSink().toString()
-                if (logLine && logLine.contains(probe.logContains)) {
+                if (logLine && logLine.contains(livenessProbe.logContains)) {
                     matchFound = true
                     break
                 } else {
-                    progressLogger.progress(sprintf('probing for %010dms', pollTimes * probe.pollInterval))
+
+                    long totalMillis = pollTimes * livenessProbe.pollInterval
+                    long totalMinutes = TimeUnit.MILLISECONDS.toMinutes(totalMillis)
+                    progressLogger.progress("Probing for ${totalMinutes}m...")
                     try {
 
                         // zero'ing out the below so as to save on memory for potentially
@@ -91,8 +107,8 @@ class DockerLivenessProbeContainer extends DockerLogsContainer {
                         logLine = null
                         getSink().getBuffer().setLength(0)
 
-                        localPollTime = localPollTime - probe.pollInterval
-                        sleep(probe.pollInterval)
+                        localPollTime -= livenessProbe.pollInterval
+                        sleep(livenessProbe.pollInterval)
                     } catch (Exception e) {
                         throw e
                     }
@@ -101,36 +117,41 @@ class DockerLivenessProbeContainer extends DockerLogsContainer {
             progressLogger.completed()
 
             if (!matchFound) {
-                throw new GradleException("Liveness probe failed to find a match: ${probe.toString()}")
+                throw new GradleException("Liveness probe failed to find a match: ${livenessProbe.toString()}")
             }
         } else {
-            def container = dockerClient.inspectContainerCmd(getContainerId()).exec()
-            if (container.getState().getRunning() == false) {
+            lastInspection = dockerClient.inspectContainerCmd(getContainerId()).exec()
+            if (lastInspection.getState().getRunning() == false) {
                 throw new GradleException("Container with ID '${getContainerId()}' is not running.");
             }
         }
     }
 
     /**
-     * Define the probe options for this liveness check. We'll default to
-     * probing for 10 minutes with 30 second intervals between each probe.
+     * Define the livenessProbe options for this liveness check. We'll default to
+     * probing for 10 minutes with 30 second intervals between each livenessProbe.
      *
      * @param logContains content within container log we will search for
      * @return instance of LivenessProbe
      */
-    def probe(final String logContains) {
-        probe(600000, 30000, logContains)
+    def livenessProbe(final String logContains) {
+        livenessProbe(600000, 30000, logContains)
     }
 
     /**
-     * Define the probe options for this liveness check.
+     * Define the livenessProbe options for this liveness check.
      *
      * @param pollTime how long we will poll for
      * @param pollInterval interval between poll requests
      * @param logContains content within container log we will search for
      * @return instance of LivenessProbe
      */
-    def probe(final long pollTime, final long pollInterval, final String logContains) {
-        this.probe = new LivenessProbe(pollTime, pollInterval, logContains)
+    def livenessProbe(final long pollTime, final long pollInterval, final String logContains) {
+        this.livenessProbe = new LivenessProbe(pollTime, pollInterval, logContains)
+    }
+
+    // return the last inspection made during the lifetime of this task.
+    def lastInspection() {
+        lastInspection
     }
 }